feat: exported method to get cookie config
